+++ lib/widgets/app_badge.dart
@@ -12,7 +12,7 @@ class AppBadge extends StatelessWidget {
       child: Container(
         padding: EdgeInsets.symmetric(vertical: 8, horizontal: 10),
         decoration: BoxDecoration(
-            borderRadius: BorderRadius.circular(10),
+            borderRadius: BorderRadius.circular(20),
             border:
                 Border.all(width: 0.5, color: Theme.of(context).primaryColor),
             color: Colors.white.withOpacity(0.9),
@@ -20,8 +20,8 @@ class AppBadge extends StatelessWidget {
               BoxShadow(
                 color: Colors.grey.withOpacity(0.3),
                 spreadRadius: 2,
-                blurRadius: 3,
-                offset: Offset(0, 3), // changes position of shadow
+                blurRadius: 2,
+                offset: Offset(0, 0), // changes position of shadow
               ),
             ]),
         child: Text(